feat: modify DTO
