--- app/src/main/java/campus/tech/kakao/map/Document.kt
@@ -1,16 +1,0 @@
-package campus.tech.kakao.map
-
-data class Document(
-    val address_name: String,
-    val category_group_code: String,
-    val category_group_name: String,
-    val category_name: String,
-    val distance: String,
-    val id: String,
-    val phone: String,
-    val place_name: String,
-    val place_url: String,
-    val road_address_name: String,
-    val x: String,
-    val y: String
-)
