--- app/src/main/java/campus/tech/kakao/map/KaKaoLocalData.kt
@@ -1,6 +1,0 @@
-package campus.tech.kakao.map
-
-data class KaKaoLocalData(
-    val documents: List<Document>,
-    val meta: Meta
-)
