+++ app/src/main/java/campus/tech/kakao/map/KakaoAPIRetrofitService.kt
@@ -7,8 +7,9 @@ import retrofit2.http.Query
 
 interface KakaoAPIRetrofitService {
     @GET("v2/local/search/keyword.json")
+
     fun getSearchKeyword(
         @Header("Authorization") key: String,
         @Query("query") query: String
-    ): Call<KaKaoLocalData>
+    ): Call<SearchResponse>
 }
