--- app/src/main/java/campus/tech/kakao/map/Meta.kt
@@ -1,8 +1,0 @@
-package campus.tech.kakao.map
-
-data class Meta(
-    val is_end: Boolean,
-    val pageable_count: Int,
-    val same_name: SameName,
-    val total_count: Int
-)
