--- app/src/main/java/campus/tech/kakao/map/SameName.kt
@@ -1,7 +1,0 @@
-package campus.tech.kakao.map
-
-data class SameName(
-    val keyword: String,
-    val region: List<Any>,
-    val selected_region: String
-)
